% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cli_version}
\alias{cli_version}
\title{Get the daggle CLI version}
\usage{
cli_version()
}
\value{
A character string with the daggle CLI version.
}
\description{
Shells out to \verb{daggle version} and returns the version string.
}
\examples{
\dontrun{
cli_version()
}
}
