% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{compare_runs}
\alias{compare_runs}
\title{Compare two runs}
\usage{
compare_runs(name, run1, run2, base_url = NULL)
}
\arguments{
\item{name}{Character string. Name of the DAG.}

\item{run1}{Character string. First run ID.}

\item{run2}{Character string. Second run ID.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A list with elements: \code{outputs_diff} (data.frame with columns
\code{step_id}, \code{key}, \code{value1}, \code{value2}), \code{duration_diff} (list with
\code{run1_seconds}, \code{run2_seconds}, \code{diff_seconds}), \code{meta_diff} (list with
\code{dag_hash1}, \code{dag_hash2}, \code{changed}).
}
\description{
Compare two runs
}
