% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{daggle_approve}
\alias{daggle_approve}
\title{Approve a waiting step}
\usage{
daggle_approve(name, run_id, step_id, base_url = NULL)
}
\arguments{
\item{name}{Character string. Name of the DAG.}

\item{run_id}{Character string. Run ID.}

\item{step_id}{Character string. Step ID to approve.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A list with elements: \code{step_id}, \code{status}.
}
\description{
Approve a waiting step
}
