% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{daggle_cancel_run}
\alias{daggle_cancel_run}
\title{Cancel a running DAG run}
\usage{
daggle_cancel_run(name, run_id = "latest", base_url = NULL)
}
\arguments{
\item{name}{Character string. Name of the DAG.}

\item{run_id}{Character string. Run ID, or \code{"latest"} for the most recent.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A list with elements: \code{status}, \code{run_id}, \code{message}.
}
\description{
Cancel a running DAG run
}
