% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-step.R
\name{daggle_dag_name}
\alias{daggle_dag_name}
\title{Get the current DAG name}
\usage{
daggle_dag_name()
}
\value{
Character string from the \code{DAGGLE_DAG_NAME} environment variable.
}
\description{
Get the current DAG name
}
\examples{
\dontrun{
daggle_dag_name()
}
}
