% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{daggle_health}
\alias{daggle_health}
\title{Check API health}
\usage{
daggle_health(base_url = NULL)
}
\arguments{
\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A list with elements: \code{status}, \code{version}, \code{uptime_seconds}.
}
\description{
Check API health
}
