% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-step.R
\name{daggle_meta_text}
\alias{daggle_meta_text}
\title{Emit a text metadata value}
\usage{
daggle_meta_text(name, value)
}
\arguments{
\item{name}{Character string. Metadata key. Must match \verb{[a-zA-Z_][a-zA-Z0-9_]*}.}

\item{value}{Character string. Metadata value.}
}
\value{
\code{value}, invisibly.
}
\description{
Emit a text metadata value
}
\examples{
\dontrun{
daggle_meta_text("model_type", "linear regression")
}
}
