% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-step.R
\name{daggle_run_dir}
\alias{daggle_run_dir}
\title{Get the current run directory}
\usage{
daggle_run_dir()
}
\value{
Character string from the \code{DAGGLE_RUN_DIR} environment variable.
}
\description{
Get the current run directory
}
\examples{
\dontrun{
daggle_run_dir()
}
}
