% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-step.R
\name{daggle_run_id}
\alias{daggle_run_id}
\title{Get the current daggle run ID}
\usage{
daggle_run_id()
}
\value{
Character string from the \code{DAGGLE_RUN_ID} environment variable.
}
\description{
Get the current daggle run ID
}
\examples{
\dontrun{
daggle_run_id()
}
}
