% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_metadata}
\alias{get_metadata}
\title{Get step metadata for a run}
\usage{
get_metadata(name, run_id = "latest", base_url = NULL)
}
\arguments{
\item{name}{Character string. Name of the DAG.}

\item{run_id}{Character string. Run ID, or \code{"latest"} for the most recent.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A data.frame with columns: \code{step_id}, \code{name}, \code{type}, \code{value}.
}
\description{
Get step metadata for a run
}
