% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_step_log}
\alias{get_step_log}
\title{Get log output for a step}
\usage{
get_step_log(name, run_id, step_id, base_url = NULL)
}
\arguments{
\item{name}{Character string. Name of the DAG.}

\item{run_id}{Character string. Run ID.}

\item{step_id}{Character string. Step ID.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A list with elements: \code{step_id}, \code{stdout}, \code{stderr}.
}
\description{
Get log output for a step
}
