% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_projects}
\alias{list_projects}
\title{List registered projects}
\usage{
list_projects(base_url = NULL)
}
\arguments{
\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A data.frame with columns: \code{name}, \code{path}, \code{status}, \code{dags}.
}
\description{
List registered projects
}
