% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{list_runs}
\alias{list_runs}
\title{List runs for a DAG}
\usage{
list_runs(name, base_url = NULL)
}
\arguments{
\item{name}{Character string. Name of the DAG.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A data.frame with columns: \code{run_id}, \code{started}, \code{status},
\code{duration_seconds}, \code{dag_hash}.
}
\description{
List runs for a DAG
}
