% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{register_project}
\alias{register_project}
\title{Register a project}
\usage{
register_project(path, name = NULL, base_url = NULL)
}
\arguments{
\item{path}{Character string. Absolute path to the project directory.}

\item{name}{Character string or \code{NULL}. Optional project name; defaults to
the directory basename on the server side.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A list with elements: \code{name}, \code{path}.
}
\description{
Register a project
}
