% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-step.R
\name{run_dir}
\alias{run_dir}
\title{Get the current run directory}
\usage{
run_dir()
}
\value{
Character string from the \code{DAGGLE_RUN_DIR} environment variable.
}
\description{
Get the current run directory
}
\examples{
\dontrun{
run_dir()
}
}
