% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in-step.R
\name{run_id}
\alias{run_id}
\title{Get the current daggle run ID}
\usage{
run_id()
}
\value{
Character string from the \code{DAGGLE_RUN_ID} environment variable.
}
\description{
Get the current daggle run ID
}
\examples{
\dontrun{
run_id()
}
}
