% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{trigger}
\alias{trigger}
\title{Trigger a new DAG run}
\usage{
trigger(name, params = list(), base_url = NULL)
}
\arguments{
\item{name}{Character string. Name of the DAG to trigger.}

\item{params}{Named list of parameters to pass to the run.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A list with elements: \code{run_id}, \code{status}.
}
\description{
Trigger a new DAG run
}
