% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{unregister_project}
\alias{unregister_project}
\title{Unregister a project}
\usage{
unregister_project(name, base_url = NULL)
}
\arguments{
\item{name}{Character string. Name of the project to unregister.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A list with element: \code{name}.
}
\description{
Unregister a project
}
