% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_dag}
\alias{get_dag}
\title{Get details for a single DAG}
\usage{
get_dag(name, base_url = NULL)
}
\arguments{
\item{name}{Character string. Name of the DAG.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A list with elements: \code{name}, \code{steps}, \code{step_ids}, \code{schedule},
\code{workdir}, \code{r_version}, \code{last_status}, \code{last_run_id}, \code{last_run}.
}
\description{
Get details for a single DAG
}
