% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_run}
\alias{get_run}
\title{Get details for a specific run}
\usage{
get_run(name, run_id = "latest", base_url = NULL)
}
\arguments{
\item{name}{Character string. Name of the DAG.}

\item{run_id}{Character string. Run ID, or \code{"latest"} for the most recent.}

\item{base_url}{Base URL of the daggle API. See Details.}
}
\value{
A list with elements: \code{run_id}, \code{dag_name}, \code{status}, \code{started},
\code{ended}, \code{duration_seconds}, \code{dag_hash}, \code{r_version}, \code{platform},
\code{params}, \code{steps}.
}
\description{
Get details for a specific run
}
